+++ lib/core/router/router_paths.dart
@@ -12,6 +12,9 @@ abstract class AppRoutePaths {
   /// Route path for the sign-up page.
   static const signUpPath = '$authPrefix/sign-up';
 
+  /// Route path for the forgot-password page.
+  static const forgotPasswordPath = '$authPrefix/forgot-password';
+
   /// Route path for the verify-email page.
   static const verifyEmailPath = '$authPrefix/verify-email';
 }